+++ packages/atomic/src/components/search/result-template-components/quickview-iframe/quickview-iframe.tsx
@@ -1,4 +1,4 @@
-import {Result} from '@coveo/headless';
+import {SearchEngine} from '@coveo/headless';
 import {FunctionalComponent, h} from '@stencil/core';
 
 const documentIdentifierInIframe = 'CoveoDocIdentifier';
@@ -14,35 +14,44 @@ const writeDocument = (documentWriter: Document, content: string) => {
 
 const currentResultAlreadyWrittenToDocument = (
   documentWriter: Document,
-  result: Result
+  uniqueIdentifier: string
 ) => {
   const currentDocIdentifier = documentWriter.getElementById(
     documentIdentifierInIframe
   );
 
   return (
-    currentDocIdentifier && currentDocIdentifier.textContent === result.uniqueId
+    currentDocIdentifier &&
+    currentDocIdentifier.textContent === uniqueIdentifier
   );
 };
 
 const ensureSameResultIsNotOverwritten = (
   documentWriter: Document,
-  result: Result
+  uniqueIdentifier: string
 ) => {
   const docIdentifier = documentWriter.createElement('div');
   docIdentifier.style.display = 'none';
   docIdentifier.setAttribute('aria-hidden', 'true');
   docIdentifier.id = documentIdentifierInIframe;
-  docIdentifier.textContent = result.uniqueId;
+  docIdentifier.textContent = uniqueIdentifier;
   documentWriter.body.appendChild(docIdentifier);
+};
+
+const warnAboutLimitedUsageQuickview = (logger?: SearchEngine['logger']) => {
+  logger?.warn(
+    'Quickview initialized in restricted mode due to incompatible sandboxing environment. Keywords hit navigation will be disabled.'
+  );
 };
 
 export const QuickviewIframe: FunctionalComponent<{
   content?: string;
   onSetIframeRef: (ref: HTMLIFrameElement) => void;
-  result?: Result;
+  uniqueIdentifier?: string;
   sandbox?: string;
-}> = ({onSetIframeRef, result, content, sandbox}) => {
+  src?: string;
+  logger?: SearchEngine['logger'];
+}> = ({onSetIframeRef, uniqueIdentifier, content, sandbox, src, logger}) => {
   // When a document is written with document.open/document.write/document.close
   // it is not synchronous and the content of the iframe is only available to be queried at the end of the current call stack.
   // This add a "wait" (setTimeout 0) before calling the `onSetIframeRef` from the parent modal quickview
@@ -58,20 +67,30 @@ export const QuickviewIframe: FunctionalComponent<{
       ref={async (el) => {
         const iframeRef = el as HTMLIFrameElement;
 
-        if (!result || !content) {
+        if (!uniqueIdentifier || !content) {
           return;
         }
 
         const documentWriter = iframeRef.contentDocument;
         if (!documentWriter) {
+          if (src) {
+            warnAboutLimitedUsageQuickview(logger);
+            iframeRef.src = src;
+          }
+
           return;
         }
-        if (currentResultAlreadyWrittenToDocument(documentWriter, result)) {
+        if (
+          currentResultAlreadyWrittenToDocument(
+            documentWriter,
+            uniqueIdentifier
+          )
+        ) {
           return;
         }
 
         writeDocument(documentWriter, content);
-        ensureSameResultIsNotOverwritten(documentWriter, result);
+        ensureSameResultIsNotOverwritten(documentWriter, uniqueIdentifier);
 
         await waitForIframeContentToBeWritten();
         onSetIframeRef(iframeRef);